deleted drop-down menu page
